+++ .github/CHANGELOG.md
@@ -1,5 +1,6 @@
 # March 2023
 - feat: (03/07/2023): Integrate slash command api & new cowsay command
+- feat: (03/27/2023): Integrate OpenAI API & ChatGPT Command
 
 # November 2022
 - feat: (11/04/2022): Initialize new discord.js bot